Create Recursive menu of Permutation, Combination and Fibonacci
nPr (Permutation Calculation)
In the program, nPr is used to find how many ways we can arrange r items from n items when order matters. The formula is nPr = n! / (n - r)!, which means we find the factorial of n and divide it by the factorial of (n - r). The code does this by using the factorial() function, which calls itself recursively to calculate the factorial of a number.

nCr (Combination Calculation)
The nCr option calculates how many ways we can choose r items from n items when order doesn't matter. The formula is nCr = n! / (r! * (n - r)!), meaning we divide n! by both r! and (n - r)! to remove duplicate arrangements. In the code, this is also done using the factorial() function, which keeps calling itself until it reaches 1.

Fibonacci Sequence Calculation
The Fibonacci sequence is a series of numbers where each number is the sum of the two before it. The code uses the fibonacci() function, which calls itself to find Fibonacci(n) = Fibonacci(n - 1) + Fibonacci(n - 2). It stops when n is 0 or 1, returning 1. This method works but is slow for large numbers because it repeats calculations many times.